requirements.txt
amqp==5.2.0
annotated-types==0.7.0
asttokens==2.4.1
billiard==4.2.0
blinker==1.8.2
celery==5.4.0
cffi==1.17.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.6.0
cryptography==43.0.1
dnspython==2.6.1
email_validator==2.2.0
executing==2.0.1
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn==22.0.0
icecream==2.1.3
idna==3.9
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kombu==5.4.1
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
numpy==2.1.1
packaging==24.1
pandas==2.2.2
pluggy==1.5.0
prompt_toolkit==3.0.47
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
pymongo==4.8.0
pyodbc==5.1.0
pyOpenSSL==24.2.1
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.0.8
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
SQLAlchemy==2.0.31
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.3
WTForms==3.1.2
